OAUTH2-120 Move to oauth2-provider-jsonws
