--- modules/apps/oauth2-provider/oauth2-provider-json-web-service/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: Liferay OAuth2 Provider JSON Web Service
-Bundle-SymbolicName: com.liferay.oauth2.provider.json.web.service
-Bundle-Version: 1.0.0
