--- modules/apps/oauth2-provider/oauth2-provider-json-web-service/src/main/java/com/liferay/oauth2/provider/json/web/service/internal/instance/lifecycle/OAuth2JSONWebServicePortalInstanceLifecycleListener.java
@@ -1,41 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.oauth2.provider.json.web.service.internal.instance.lifecycle;
-
-import com.liferay.oauth2.provider.json.web.service.internal.service.access.policy.scope.SAPEntryScopeDescriptorFinderRegistrator;
-import com.liferay.portal.instance.lifecycle.BasePortalInstanceLifecycleListener;
-import com.liferay.portal.instance.lifecycle.PortalInstanceLifecycleListener;
-import com.liferay.portal.kernel.model.Company;
-
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-
-/**
- * @author Tomas Polesovsky
- */
-@Component(service = PortalInstanceLifecycleListener.class)
-public class OAuth2JSONWebServicePortalInstanceLifecycleListener
-	extends BasePortalInstanceLifecycleListener {
-
-	public void portalInstanceRegistered(Company company) {
-		_sapEntryScopeDescriptorFinderRegistrator.register(
-			company.getCompanyId());
-	}
-
-	@Reference
-	private SAPEntryScopeDescriptorFinderRegistrator
-		_sapEntryScopeDescriptorFinderRegistrator;
-
-}
