--- modules/apps/oauth2-provider/oauth2-provider-json-web-service/src/main/resources/META-INF/module-log4j.xml
@@ -1,7 +1,0 @@
-<!DOCTYPE log4j:configuration SYSTEM "log4j.dtd">
-
-<log4j:configuration xmlns:log4j="http://jakarta.apache.org/log4j/">
-	<category name="com.liferay.oauth2.provider.json.web.service">
-		<priority value="INFO" />
-	</category>
-</log4j:configuration>
