+++ modules/apps/oauth2-provider/oauth2-provider-jsonws/bnd.bnd
@@ -1,0 +1,3 @@
+Bundle-Name: Liferay OAuth2 Provider JSONWS
+Bundle-SymbolicName: com.liferay.oauth2.provider.jsonws
+Bundle-Version: 1.0.0
