+++ modules/apps/oauth2-provider/oauth2-provider-jsonws/build.gradle
