+++ modules/apps/oauth2-provider/oauth2-provider-jsonws/src/main/java/com/liferay/oauth2/provider/jsonws/internal/application/descriptor/OAuth2JSONWSApplicationDescriptor.java
@@ -1,0 +1,71 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.oauth2.provider.jsonws.internal.application.descriptor;
+
+import com.liferay.oauth2.provider.jsonws.internal.configuration.OAuth2JSONWSConfiguration;
+import com.liferay.oauth2.provider.jsonws.internal.constants.OAuth2JSONWSConstants;
+import com.liferay.oauth2.provider.scope.spi.application.descriptor.ApplicationDescriptor;
+import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
+import com.liferay.portal.kernel.util.ResourceBundleLoader;
+import com.liferay.portal.kernel.util.ResourceBundleUtil;
+
+import java.util.Locale;
+import java.util.Map;
+
+import org.osgi.service.component.annotations.Activate;
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author Tomas Polesovsky
+ */
+@Component(
+	configurationPid = "com.liferay.oauth2.provider.jsonws.internal.configuration.OAuth2JSONWSConfiguration",
+	property = "osgi.jaxrs.name=" + OAuth2JSONWSConstants.LIFERAY_JSON_WEB_SERVICES
+)
+public class OAuth2JSONWSApplicationDescriptor
+	implements ApplicationDescriptor {
+
+	@Override
+	public String describeApplication(Locale locale) {
+		String applicationDescription = ResourceBundleUtil.getString(
+			_resourceBundleLoader.loadResourceBundle(locale),
+			_applicationDescription);
+
+		if (applicationDescription == null) {
+			return _applicationDescription;
+		}
+
+		return applicationDescription;
+	}
+
+	@Activate
+	protected void activate(Map<String, Object> properties) {
+		OAuth2JSONWSConfiguration oAuth2JSONWSConfiguration =
+			ConfigurableUtil.createConfigurable(
+				OAuth2JSONWSConfiguration.class, properties);
+
+		_applicationDescription =
+			oAuth2JSONWSConfiguration.applicationDescription();
+	}
+
+	private String _applicationDescription;
+
+	@Reference(
+		target = "(bundle.symbolic.name=com.liferay.oauth2.provider.jsonws)"
+	)
+	private ResourceBundleLoader _resourceBundleLoader;
+
+}
