+++ modules/apps/oauth2-provider/oauth2-provider-jsonws/src/main/resources/content/Language.properties
@@ -1,0 +1,11 @@
+OAUTH2_everything=Grants all access to all portal remote JSONWS services
+OAUTH2_everything.readonly=Grants READ-ONLY access to all portal remote JSONWS services
+OAUTH2_userprofile=Read your personal user data
+oauth2.application.description.JSONWS=Portal JSONWS API Service Access Policies
+oauth2-jsonws-configuration-name=OAuth2 JSON Web Service Configuration
+oauth2-jsonws-application-description=OAuth2 JSON Web Service Application Description
+oauth2-jsonws-application-description-description=Localization key or string to describe Portal JSONWS API OAuth2 Application
+oauth2-remove-sap-entry-oauth2-prefix=Remove OAuth2 Service Access Policy Prefix
+oauth2-remove-sap-entry-oauth2-prefix-description=When unselected, SAP entry name will be used as-is for scope name
+oauth2-sap-entry-oauth2-prefix=OAuth2 Service Access Policy Prefix
+oauth2-sap-entry-oauth2-prefix-description=Prefix for Service Access Policies to be included in OAuth2 framework
